Move graph components into a sub directory
