+++ src/components/graphs/BarGraph.vue
